refactor: removed the explicit validator package and moved all of them into helpers
