--- server/validators/mime_type_validators.go
@@ -1,30 +1,0 @@
-package validators
-
-import (
-	"fmt"
-	"regexp"
-	"strings"
-)
-
-func ValidateMimeType(mimeType string) bool {
-	mimeTypePattern := `^[a-zA-Z]+/[a-zA-Z0-9\-\.\+]+$`
-
-	re := regexp.MustCompile(mimeTypePattern)
-
-	return re.MatchString(mimeType)
-}
-
-func ValidateAllowedMimeTypes(mimeTypes []string) error {
-	var invalidContentTypes []string
-	for _, mimeType := range mimeTypes {
-		if !ValidateMimeType(mimeType) {
-			invalidContentTypes = append(invalidContentTypes, mimeType)
-		}
-	}
-
-	if len(invalidContentTypes) > 0 {
-		return fmt.Errorf("invalid content types: [%s]", strings.Join(invalidContentTypes, ", "))
-	}
-
-	return nil
-}
