--- server/validators/string_validators.go
@@ -1,11 +1,0 @@
-package validators
-
-import "strings"
-
-func ValidateNotEmptyTrimmedString(value string) bool {
-	if strings.Trim(value, " ") == "" {
-		return true
-	}
-
-	return false
-}
